refactor: BettingResultAnalyzer 기능을 GameResultAnalyzer로 이동하고 승패판정 기능 중복 삭제
